Bug fix for MFSim target:
details: when a non-split droplet has multiple successors (this occurs anytime we detect a droplet, as detect does not 'consume' this droplet), we previously drew edges to all successors.  We now find the use(s) that does not rename the droplet (currently only a detect will cause this) and only draw edges to those uses.

This enforces the following semantic:  each droplet (in the generated MFSim DAGs) is defined and used at most once.